chg: adapt printout: total `clicks`/`impressions`
The printed messages declare `impression`/`clicks` metrics to be *daily* impressions/clicks. This does not appear to be correct: When creating a keyword plan with the code example in `./add_keyword_plan.py` and generating metrics for that plan, the output values scale linearly with the timespan used in the `add_keyword_plan._create_keyword_plan` - i.e. roughly 4 times higher for `NEXT_MONTH` than for  `NEXT_WEEK`, roughly 3 times higher still for `NEXT_QUARTER`